dash/requirements.txt
pandas
psycopg2-binary
sqlalchemy
dash
plotly
dash-bootstrap-components
streamlit